from_mqtt/mqtt_sql/requirements.txt
paho-mqtt==2.1.0
PyYAML==6.0.1
logging==0.4.9.6
pymssql==2.3.0